[all components] Rename Vitest matcher setup
