+++ packages/react/tsconfig.test.json
@@ -17,7 +17,7 @@
     "src/**/*.test.ts*",
     "src/index.test.ts",
     "test",
-    "test/addChaiAssertions.ts",
+    "test/addVitestMatchers.ts",
     "src/global.d.ts",
     "package.json"
   ],
